--update master MergeRealty

+++ js/component/NewContact.js
@@ -218,11 +218,12 @@ NewContact.prototype.getDataSave = function() {
         statusphone: this.type.value,
         email: this.email.value
     }
-    if (moduleDatabase.getModule("users").getLibary("phone")[this.phone.value] !== undefined || moduleDatabase.getModule("contacts").getLibary("phone")[this.phone.value] !== undefined) {
-        var deleteItem = confirmQuestion("Xác nhận số điện thoại", "Số điện thoại trùng vui lòng kiểm tra lại!");
-        this.$view.addChild(deleteItem);
-        return;
-    }
+    if (this.data == undefined)
+        if (moduleDatabase.getModule("users").getLibary("phone")[this.phone.value] !== undefined || moduleDatabase.getModule("contacts").getLibary("phone")[this.phone.value] !== undefined) {
+            var deleteItem = confirmQuestion("Xác nhận số điện thoại", "Số điện thoại trùng vui lòng kiểm tra lại!");
+            this.$view.addChild(deleteItem);
+            return;
+        }
     if (this.data !== undefined)
         temp.id = this.data.original.id;
     return temp;
